Update Model & Controller

+++ src/controllers/image.ts
@@ -3,7 +3,7 @@ import Image from '../models/image'
 import { APP_URL } from "../config/config"
 import path from 'path'
 import fs from 'fs'
-import slugify from 'slugify'
+
 
 
 const getAllImage = async (req: Request, res: Response) => {
@@ -11,19 +11,19 @@ const getAllImage = async (req: Request, res: Response) => {
         const images = await Image.find({})
         // .select({ _id: 0, __v: 0})
 
-        let imagesDummy = [];
+        // let imagesDummy = [];
 
-        for await (let image of images) {
-            if (image.imageURL) {
-                image.imageURL = APP_URL + image.imageURL;
-            }
-            imagesDummy.push(image);
-        }
+        // for await (let image of images) {
+        //     if (image.imageURL) {
+        //         image.imageURL = APP_URL + image.imageURL;
+        //     }
+        //     imagesDummy.push(image);
+        // }
 
-        console.log('images', images);
+        // console.log('images', images);
 
 
-        return res.status(200).json({ error: false, data: imagesDummy })
+        return res.status(200).json({ error: false, data: images })
     } catch (error) {
         console.log(error);
         return res.status(200).json({ error: true, msg: "Server Error" })
@@ -41,91 +41,8 @@ const getImage = async (req: Request, res: Response) => {
         const image = await Image.findById({ _id: imageId }
         ).select({ _id: 0, __v: 0 })
 
-
-        // Image.aggregate([
-        //     {
-        //         "$match": {
-        //             _id: '612dfe2e95b1b329649c698d'
-        //         }
-        //     },
-        //     // { $match: {  _id: imageId} },
-        //     {
-        //         "$set": {
-        //             "imageURL": {
-        //                 "$concat": [
-        //                     APP_URL,
-        //                     "$imageURL"
-        //                 ]
-        //             }
-        //         }
-        //     }
-        // ]).then(result => {
-        //     console.log(result);
-
-        //     return res.status(200).json({
-        //         success: true,
-        //         data: result
-        //     })
-        // }).catch(error => {
-        //     console.log(error);
-
-        //     return res.status(400).json({
-        //         success: false,
-        //         error: error
-        //     })
-        // })
-
-
         return res.status(200).json({ error: false, data:image  })
 
-    } catch (error) {
-        console.log(error);
-        return res.status(200).json({ error: true, msg: "Server Error" })
-    }
-}
-
-
-const getImageBySlug = async (req: Request, res: Response) => {
-    try {
-        const imageSlug = req.params.slug
-        const findImage = await Image.findOne({ slug: imageSlug })
-        if (!findImage) {
-            return res.status(404).json({ error: true, msg: "Image not found" })
-        }
-        const image = await Image.findById({ slug: imageSlug }).select({ _id: 0, __v: 0 })
-        // Image.aggregate([
-        //     {
-        //         "$match": {
-        //             _id: imageSlug
-        //         }
-        //     },
-        //     // { $match: {  _id: imageId} },
-        //     {
-        //         "$set": {
-        //             "imageURL": {
-        //                 "$concat": [
-        //                     APP_URL,
-        //                     "$imageURL"
-        //                 ]
-        //             }
-        //         }
-        //     }
-        // ]).then(result => {
-        //     console.log(result);
-
-        //     return res.status(200).json({
-        //         success: true,
-        //         data: result
-        //     })
-        // }).catch(error => {
-        //     console.log(error);
-
-        //     return res.status(400).json({
-        //         success: false,
-        //         error: error
-        //     })
-        // })
-        return res.status(200).json({ error: false, data: image })
     } catch (error) {
         console.log(error);
         return res.status(200).json({ error: true, msg: "Server Error" })
@@ -138,7 +55,6 @@ const addImage = async (req: Request, res: Response) => {
         if (req.file) {
             console.log(req.file)
 
-
         }
 
         if (!req.file) {
@@ -148,25 +64,16 @@ const addImage = async (req: Request, res: Response) => {
         const file = req.file.filename
         const ext = path.extname(file).split(".");
         const file_type = ext[1]
-        // console.log('file_type', file_type);
+ 
         const file_name = file.replace(file_type, "").split('.')
 
         const image_title = file_name[0]
         console.log("image_title", image_title)
 
-        let slug = slugify(image_title, { lower: true });
-
-        let slug_exist = await Image.findOne({
-            slug: slug
-        });
-        if (slug_exist) {
-            slug = slug + "-" + Math.floor(Math.random() * 90 + 10);
-        }
-
         const image = new Image({
             title: image_title,
-            slug: slug,
-            imageURL: req.file.path
+            imageURL: req.file.path,
+            imageFullURL: APP_URL + req.file.path
         })
         await image.save()
         return res.status(201).json({ error: false, msg: "Add Succesfully", data: image })
@@ -176,41 +83,26 @@ const addImage = async (req: Request, res: Response) => {
     }
 }
 
-// const addImageByLink = async (req: Request, res: Response) => {
-//     try {
+const addImageByLink = async (req: Request, res: Response) => {
+    try {
 
-        
+        let {imageFullURL} = req.body
 
-//         const file = req.file.filename
-//         const ext = path.extname(file).split(".");
-//         const file_type = ext[1]
-//         // console.log('file_type', file_type);
-//         const file_name = file.replace(file_type, "").split('.')
+        if(imageFullURL== undefined || imageFullURL== null || imageFullURL== ''){
+            return res.status(401).json({ error: true, msg: "Image URL requried" })
+        }
 
-//         const image_title = file_name[0]
-//         console.log("image_title", image_title)
+        const image = new Image({
+            imageFullURL,
+        })
+        await image.save()
+        return res.status(201).json({ error: false, msg: "Add Succesfully", data: image })
+    } catch (error) {
+        console.log(error);
+        return res.status(200).json({ error: true, msg: "Server Error" })
+    }
+}   
 
-//         let slug = slugify(image_title, { lower: true });
-
-//         let slug_exist = await Image.findOne({
-//             slug: slug
-//         });
-//         if (slug_exist) {
-//             slug = slug + "-" + Math.floor(Math.random() * 90 + 10);
-//         }
-
-//         const image = new Image({
-//             title: image_title,
-//             slug: slug,
-//             imageURL: req.file.path
-//         })
-//         await image.save()
-//         return res.status(201).json({ error: false, msg: "Add Succesfully", data: image })
-//     } catch (error) {
-//         console.log(error);
-//         return res.status(200).json({ error: true, msg: "Server Error" })
-//     }
-// }
 
 
 const removeImage = async (req: Request, res: Response) => {
@@ -221,8 +113,10 @@ const removeImage = async (req: Request, res: Response) => {
             return res.status(404).json({ error: true, msg: "Image not found" })
         }
         const image = await Image.findByIdAndRemove({ _id: imageId })
-
+        
+        if(image?.imageURL){
         fs.unlinkSync(`${image?.imageURL}`);
+        }
 
         return res.status(200).json({ error: false, msg: "Delete Successfuly" })
     } catch (error) {
@@ -236,6 +130,6 @@ export default {
     getAllImage,
     getImage,
     removeImage,
-    getImageBySlug,
     addImage,
+    addImageByLink
 }
